Starting to refactor module structure
