+++ old_new_structure.md
@@ -69,7 +69,7 @@ mbs_results/
 │   └── winsorisation.py
 ├── outputs/ - *New Proposed Change, move outputs from branch to main package*
 │   ├── `__init__.py`
-│   └── growth_rate_output.py - Luke, where to go 
+│   └── growth_rate_output.py
 ├── utilities/
 │   ├── utils.py
 │   ├── data_cleaning.py
@@ -79,3 +79,16 @@ mbs_results/
     ├── selective_editing.py?
     ├── turnover_analysis.py
     └── validation_checks.py
+
+test/
+|   test_imputation/
+│   ├── `__init__.py`
+│   ├── calculate_imputation_link.py
+│   ├── cumulative_imputation_links.py
+│   ├── flag_and_count_matched_pairs.py
+│   ├── imputation_flags.py
+│   ├── link_filter.py
+│   ├── pivot_imputation_value.py
+│   ├── predictive_variable.py
+│   ├── ratio_of_means.py
+│   └── validate_imputation.py
